Add files via upload
(Previous Commit was the old file)
Some bug fixes when pulling data to be classified (testing)
Added saving testing images to disk (both original and resized) for cross verification. 
Added an output swc file to view classified testing results. (As of now, add a segment to the file if it is not classified as negative)